requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-flask==1.2.0
aiohttp==3.8.5
python-telegram-bot==20.6
redis==4.6.0
APScheduler==3.10.4

